move create project to a dialog
